requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
python-dateutil>=2.8.2
websocket-client>=1.7.0
httpx[http2]>=0.26.0
flask-caching>=2.1.0
structlog>=24.1.0
psutil>=5.9.0
gunicorn[eventlet]>=21.2.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-dotenv>=1.0.0
email-validator>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
python-telegram-bot>=20.7
sentry-sdk[flask]>=1.30.0
alembic>=1.13.0
celery[redis]>=5.3.6
flower>=2.0.1
flask-limiter>=3.5.0
bcrypt>=4.1.2
pyotp>=2.9.0
flask-jwt-extended>=4.6.0
scipy>=1.11.0
matplotlib>=3.8.0